Rename instrumentation -> instr.
